# ideation-openapi-spec.yml
openapi: 3.0.3
info:
  title: Ideation Platform API
  version: 1.2.0
  description: |
    Comprehensive API documentation for the Ideation Platform - a NestJS-based backend for innovation management.
    
    Features:
    - JWT authentication with role-based access control
    - Idea submission with file upload support
    - Voting system with upvote/downvote functionality
    - Comment system with user associations
    - Activity logging and audit trails
    - Admin dashboard with detailed analytics
    - Real-time metrics and progress tracking
    
    All idea endpoints return IdeaResponseDto objects containing computed fields like vote counts, comment counts, and user information.
servers:
  - url: https://backendtesting-production.up.railway.app/
paths:
...
  /user/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employeeId, password]
              properties:
                employeeId:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
  /user/me:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile without innovation metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /user/{id}/activity:
    get:
      summary: Get user activity stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User activity
          content:
            application/json:
              schema:
                type: object
                properties:
                  ideasSubmitted:
                    type: integer
...
  /idea:
    post:
      summary: Create a new idea
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IdeaCreateRequest'
      responses:
        '201':
          description: Idea created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Get all ideas with filtering and pagination
      parameters:
        - name: employeeId
          in: query
          schema:
            type: string
          description: Filter by employee ID
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [recent, popular]
            default: recent
          description: Sort order
        - name: search
          in: query
          schema:
            type: string
          description: Search in idea title and description
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of records to skip for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Maximum number of records to return (max 100)
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, implemented]
          description: Filter by idea status
      responses:
        '200':
          description: List of ideas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Idea'
  /idea/{id}:
    get:
      summary: Get an idea by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Idea ID
      responses:
        '200':
          description: Idea details with comment count and vote counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '404':
          description: Idea not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update an idea
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Idea ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdeaUpdateRequest'
      responses:
        '200':
          description: Idea updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not owner of the idea
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Idea not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete an idea
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Idea ID
      responses:
        '204':
          description: Idea deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not owner of the idea
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Idea not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /idea/{id}/approve:
    patch:
      summary: Approve an idea (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Idea ID
      responses:
        '200':
          description: Idea approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Idea not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /idea/{id}/reject:
    patch:
      summary: Reject an idea (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Idea ID
      responses:
        '200':
          description: Idea rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Idea not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/activity/idea/{ideaId}:
    get:
      summary: Admin - get activities by idea
      security:
        - bearerAuth: []
      parameters:
        - name: ideaId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of activities to return
      responses:
        '200':
          description: List of activities for specific idea
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        name:
          type: string
          description: Full name of the user
        employeeId:
          type: string
          description: Employee identification number
        email:
          type: string
          format: email
          description: User's email address
        role:
          type: string
          enum: [admin, employee]
          description: User's role in the system
        gender:
          type: string
          enum: [male, female, other]
          description: User's gender
      required: [id, name, employeeId, email, role, gender]
    Idea:
      type: object
      properties:
        id:
          type: integer
          description: Unique idea identifier
        title:
          type: string
          description: Title of the idea
        description:
          type: string
          description: Detailed description of the idea
        category:
          type: string
          description: Category of the idea
        impactLevel:
          type: string
          description: Expected impact level of the idea
        hashtags:
          type: array
          items:
            type: string
          description: Array of hashtags associated with the idea
        attachmentUrls:
          type: array
          items:
            type: string
          description: URLs of uploaded attachments
        requiredResources:
          type: string
          description: Resources required to implement the idea
        anonymousId:
          type: string
          description: Anonymous identifier for the idea submitter
        status:
          type: string
          enum: [pending, approved, implmented, rejected]
          description: Current status of the idea
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the idea was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the idea was last updated
        commentCount:
          type: integer
          minimum: 0
          description: Number of comments on the idea
        upVotes:
          type: integer
          minimum: 0
          description: Number of upvotes received
        downVotes:
          type: integer
          minimum: 0
          description: Number of downvotes received
        user:
          $ref: '#/components/schemas/User'
          description: User who submitted the idea (without password)
      required: [id, title, description, category, impactLevel, hashtags, attachmentUrls, requiredResources, anonymousId, status, createdAt, updatedAt, commentCount, upVotes, downVotes, user]
    IdeaCreateRequest:
      type: object
      properties:
        title:
          type: string
          description: Title of the idea
          maxLength: 255
        description:
          type: string
          description: Detailed description of the idea
        category:
          type: string
          description: Category of the idea
        impactLevel:
          type: string
          description: Expected impact level of the idea
          enum: [low, medium, high, critical]
        hashtags:
          type: array
          items:
            type: string
          description: Array of hashtags associated with the idea
        requiredResources:
          type: string
          description: Resources required to implement the idea
        attachments:
          type: array
          items:
            type: string
            format: binary
          description: File attachments (max 5 files, 5MB each)
          maxItems: 5
      required: [title, description, category, impactLevel]
    IdeaUpdateRequest:
      type: object
      properties:
        title:
          type: string
          description: Title of the idea
          maxLength: 255
        description:
          type: string
          description: Detailed description of the idea
        category:
          type: string
          description: Category of the idea
        impactLevel:
          type: string
          description: Expected impact level of the idea
          enum: [low, medium, high, critical]
        hashtags:
          type: array
          items:
            type: string
          description: Array of hashtags associated with the idea
        requiredResources:
          type: string
          description: Resources required to implement the idea
...
    Activity:
      type: object
      properties:
        id:
          type: integer
          description: Unique activity identifier
        type:
          type: string
          enum: [idea_submitted, idea_voted, idea_commented, idea_approved, idea_rejected, idea_implemented]
          description: Type of activity performed
        userId:
          type: string
          format: uuid
          description: ID of the user who performed the activity
        ideaId:
          type: integer
          description: ID of the idea related to the activity
        metadata:
          type: object
          description: Additional metadata about the activity
          properties:
            ideaTitle:
              type: string
            voteType:
              type: string
              enum: [upvote, downvote]
            userName:
              type: string
            userEmail:
              type: string
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the activity occurred
        user:
          $ref: '#/components/schemas/User'
          description: User who performed the activity
        idea:
          $ref: '#/components/schemas/Idea'
          description: Idea related to the activity
      required: [id, type, userId, createdAt]
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message or validation errors
        error:
          type: string
          description: Error type or HTTP status name
        timestamp:
          type: string
          format: date-time
          description: ISO timestamp when the error occurred
        path:
          type: string
          description: API endpoint path where the error occurred
      required: [statusCode, message, error, timestamp, path]
...